Save and print exceptions from wait_cond()
This allows us to only print the exception stacktrace if a timeout
occurs, significantly reducing noise during tests.

To use this capability, a `wait_cond()` class must define a
`false_exceptions` attribute giving a tuple of exceptions to be treated
as `False` when raised.  If this attribute is missing, the current
behaviour (all exceptions raised by `wait_cond()` are treated as test
failures) is preserved.